refactor(inputs): extract shared input styles
